+++ src/_packages/gemini.index.ts
@@ -1,12 +1,9 @@
 // ⚠️ WARNING: This code has been generated so that any manual changes will be overwritten
-// `@promptbook/gemini`
+// `@promptbook/google`
 
 import { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION } from '../version';
-
-
 
 // Note: Exporting version from each package
 export { BOOK_LANGUAGE_VERSION, PROMPTBOOK_ENGINE_VERSION };
 
-
-// Note: Entities of the `@promptbook/gemini`
+// Note: Entities of the `@promptbook/google`