+++ src/llm-providers/google/register-constructor.ts
@@ -7,12 +7,10 @@ import { createGoogleExecutionTools } from './createGoogleExecutionTools';
  *
  * Warning: This is not useful for the end user, it is just a side effect of the mechanism that handles all available LLM tools
  *
- * @public exported from `@promptbook/gemini`
+ * @public exported from `@promptbook/google`
  * @public exported from `@promptbook/cli`
  */
-export const _GoogleRegistration: Registration = $llmToolsRegister.register(
-    createGoogleExecutionTools,
-);
+export const _GoogleRegistration: Registration = $llmToolsRegister.register(createGoogleExecutionTools);
 
 /**
  * TODO: [🎶] Naming "constructor" vs "creator" vs "factory"